Moved get_installer function to pip.utils.packaging
